graph_feature_store/requirements.txt
# requirements.txt
torch>=2.1.0
torch-geometric>=2.4.0
torch-scatter
torch-sparse

transformers>=4.35.0
sentence-transformers>=2.2.2
datasets>=2.14.0

psycopg2-binary>=2.9.0
beautifulsoup4>=4.12.0
requests>=2.31.0

grpcio>=1.59.0
grpcio-tools>=1.59.0
protobuf>=4.24.0

faiss-cpu>=1.7.4
opensearch-py>=2.3.0

xgboost>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

anthropic>=0.20.0
confluent-kafka>=2.3.0

prometheus-client>=0.18.0
matplotlib>=3.7.0
